server/requirements.txt
openai==1.5.0
langchain==0.0.351
openai==0.27.8
langchain==0.0.267
unstructured==0.11.5
tiktoken==0.5.2
pinecone-client==2.2.2
pypdf==3.17.3
sentence-transformers
python-dotenv==1.0.0
flask==2.3.2
flask-cors==4.0.0
requests==2.31.0